+++ tests/lib/rules/jsx-key.js
@@ -124,6 +124,30 @@ ruleTester.run('jsx-key', rule, {
         const onTextButtonClick = (e, item) => trackLink([, getAnalyticsUiElement(item), item.name], e);
       `,
     },
+    {
+      code: `
+        function Component({ allRatings }) {
+          return (
+            <RatingDetailsStyles>
+              {Object.entries(allRatings)?.map(([key, value], index) => {
+                const rate = value?.split(/(?=[%, /])/);
+
+                if (!rate) return null;
+
+                return (
+                  <li key={\`\${entertainment.tmdbId}\${index}\`}>
+                    <img src={\`/assets/rating/\${key}.png\`} />
+                    <span className="rating-details--rate">{rate?.[0]}</span>
+                    <span className="rating-details--rate-suffix">{rate?.[1]}</span>
+                  </li>
+                );
+              })}
+            </RatingDetailsStyles>
+          );
+        }
+      `,
+      features: ['optional chaining'],
+    },
   ]),
   invalid: parsers.all([
     {