+++ tests/lib/rules/jsx-no-constructed-context-values.js
@@ -63,9 +63,6 @@ ruleTester.run('react-no-constructed-context-values', rule, {
         }
       `,
       features: ['optional chaining'],
-      parserOptions: {
-        ecmaVersion: 2020,
-      },
     },
     {
       code: `