+++ CHANGELOG.md
@@ -12,12 +12,14 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 * [`jsx-no-literals`]: properly error on children with noAttributeStrings: true ([#3317][] @TildaDares)
 * [`jsx-key`]: catch key errors inside conditional statements ([#3320][] @TildaDares)
 * [`display-name`]: Accept forwardRef and Memo nesting in newer React versions ([#3321][] @TildaDares)
+* [`jsx-key`]: avoid a crash from optional chaining from [#3320][] ([#3327][] @ljharb)
 
 ### Changed
 * [Refactor] [`jsx-indent-props`]: improved readability of the checkNodesIndent function ([#3315][] @caroline223)
 * [Tests] [`jsx-indent`], [`jsx-one-expression-per-line`]: add passing test cases ([#3314][] @ROSSROSALES)
 * [Refactor] `boolean-prop-naming`, `jsx-indent`: avoid assigning to arguments ([#3316][] @caroline223)
 
+[#3327]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3327
 [#3321]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3321
 [#3320]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3320
 [#3317]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3317